lesson_7

+++ lesson_6/Program.cs
@@ -24,15 +24,14 @@ Console.WriteLine("Кол-во элементов > 0: " + count);
 /*
 Console.Clear();
 Console.Write("Введите число b1: ");
-int b1 = Convert.ToInt32(Console.ReadLine());
+double b1 = Convert.ToInt32(Console.ReadLine());
 Console.Write("Введите число k1: ");
-int k1 = Convert.ToInt32(Console.ReadLine());
+double k1 = Convert.ToInt32(Console.ReadLine());
 Console.Write("Введите число b2: ");
-int b2 = Convert.ToInt32(Console.ReadLine());
+double b2 = Convert.ToInt32(Console.ReadLine());
 Console.Write("Введите число k2: ");
-int k2 = Convert.ToInt32(Console.ReadLine());
+double k2 = Convert.ToInt32(Console.ReadLine());
 double x = (b2 - b1) / (k1 - k2);
 double y = k1 * x + b1;
 Console.WriteLine($"Точка пересечения {(x)} {(y)}");
-// как поправить?
 */
